pyproject.toml
[project]
name = "enc-volume-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "pydantic>=2.9.2",
    "aiohttp>=3.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
